+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `linkText` resolver.
+
 ## [1.63.0] - 2022-10-04
 
 ### Changed